refactor: reorganize managers
