+++ packages/desktop/electron/lib/managers/deep-link.manager.ts
@@ -2,7 +2,7 @@ import { ipcRenderer } from 'electron'
 
 /** Deep link manager */
 // Runs in renderer process
-export class DeepLinkManager {
+class DeepLinkManager {
     public static checkDeepLinkRequestExists(): void {
         ipcRenderer.send('check-deep-link-request-exists')
     }
