+++ packages/desktop/electron/lib/managers/index.ts
@@ -1,0 +1,4 @@
+export { default as DeepLinkManager } from './deep-link.manager'
+export { default as KeychainManager } from './keychain.manager'
+export { default as NotificationManager } from './notification.manager'
+export { default as PincodeManager } from './pincode.manager'
