+++ packages/desktop/electron/lib/managers/notification.manager.ts
