+++ packages/desktop/electron/lib/managers/pincode.manager.ts
